Migrate trace cmds to trace module
