+++ src/agent/globals.js
@@ -3,7 +3,6 @@
 /* globals */
 const pointerSize = Process.pointerSize;
 let Gcwd = '/';
-const tracehooks = {};
 const allocPool = {};
 const pendingCmds = {};
 const pendingCmdSends = [];
@@ -11,7 +10,6 @@ let sendingCommand = false;
 
 module.exports = {
   pointerSize,
-  tracehooks,
   allocPool,
   pendingCmds,
   pendingCmdSends,
