+++ src/agent/index.js
@@ -18,6 +18,7 @@ const r2 = require('./r2');
 const search = require('./search');
 const sys = require('./sys');
 const swift = require('./darwin/swift');
+const trace = require('./debug/trace');
 const utils = require('./utils');
 
 // r2->io->frida->r2pipe->r2
@@ -234,24 +235,24 @@ const commandHandlers = {
   dl: sys.dlopen,
   dlf: darwin.loadFrameworkBundle,
   'dlf-': darwin.unloadFrameworkBundle,
-  dtf: traceFormat,
-  dth: traceHook,
+  dtf: trace.traceFormat,
+  dth: trace.traceHook,
   t: types,
   't*': typesR2,
-  dt: trace,
-  dtj: traceJson,
-  dtq: traceQuiet,
-  'dt*': traceR2,
-  'dt.': traceHere,
-  'dt-': clearTrace,
-  'dt-*': clearAllTrace,
-  dtr: traceRegs,
-  dtl: traceLogDump,
-  'dtl*': traceLogDumpR2,
-  dtlq: traceLogDumpQuiet,
-  dtlj: traceLogDumpJson,
-  'dtl-': traceLogClear,
-  'dtl-*': traceLogClearAll,
+  dt: trace.trace,
+  dtj: trace.traceJson,
+  dtq: trace.traceQuiet,
+  'dt*': trace.traceR2,
+  'dt.': trace.traceHere,
+  'dt-': trace.clearTrace,
+  'dt-*': trace.clearAllTrace,
+  dtr: trace.traceRegs,
+  dtl: trace.traceLogDump,
+  'dtl*': trace.traceLogDumpR2,
+  dtlq: trace.traceLogDumpQuiet,
+  dtlj: trace.traceLogDumpJson,
+  'dtl-': trace.traceLogClear,
+  'dtl-*': trace.traceLogClearAll,
   dts: stalkTraceEverything,
   'dts?': stalkTraceEverythingHelp,
   dtsj: stalkTraceEverythingJson,
@@ -422,8 +423,6 @@ if (Process.platform === 'darwin') {
   }
 }
 
-const traceListeners = [];
-
 function radareCommandInit () {
   if (_r2) {
     return true;
@@ -569,511 +568,6 @@ function changeSelinuxContext (args) {
   return JSON.stringify({ ret: rv.value, errno: rv.errno });
 }
 
-function formatArgs (args, fmt) {
-  const a = [];
-  const dumps = [];
-  let arg; let j = 0;
-  for (let i = 0; i < fmt.length; i++, j++) {
-    try {
-      arg = args[j];
-    } catch (err) {
-      console.error('invalid format', i);
-    }
-    switch (fmt[i]) {
-      case '+':
-      case '^':
-        j--;
-        break;
-      case 'h': {
-        // hexdump pointer target, default length 128
-        // customize length with h<length>, f.e. h16 to dump 16 bytes
-        let dumpLen = 128;
-        const optionalNumStr = fmt.slice(i + 1).match(/^[0-9]*/)[0];
-        if (optionalNumStr.length > 0) {
-          i += optionalNumStr.length;
-          dumpLen = +optionalNumStr;
-        }
-        dumps.push(_hexdumpUntrusted(arg, dumpLen));
-        a.push(`dump:${dumps.length} (len=${dumpLen})`);
-      }
-        break;
-      case 'H': {
-        // hexdump pointer target, default length 128
-        // use length from other funtion arg with H<arg number>, f.e. H0 to dump '+args[0]' bytes
-        let dumpLen = 128;
-        const optionalNumStr = fmt.slice(i + 1).match(/^[0-9]*/)[0];
-        if (optionalNumStr.length > 0) {
-          i += optionalNumStr.length;
-          const posLenArg = +optionalNumStr;
-          if (posLenArg !== j) {
-            // only adjust dump length, if the length param isn't the dump address itself
-            dumpLen = +args[posLenArg];
-          }
-        }
-        // limit dumpLen, to avoid oversized dumps, caused by  accidentally parsing pointer agrs as length
-        // set length limit to 64K for now
-        const lenLimit = 0x10000;
-        dumpLen = dumpLen > lenLimit ? lenLimit : dumpLen;
-        dumps.push(_hexdumpUntrusted(arg, dumpLen));
-        a.push(`dump:${dumps.length} (len=${dumpLen})`);
-      }
-        break;
-      case 'x':
-        a.push('' + ptr(arg));
-        break;
-      case 'c':
-        a.push("'" + arg + "'");
-        break;
-      case 'i':
-        a.push(+arg);
-        break;
-      case 'z': // *s
-        const s = _readUntrustedUtf8(arg);
-        a.push(JSON.stringify(s));
-        break;
-      case 'w': // *s
-        const sw = _readUntrustedUtf16(arg);
-        a.push(JSON.stringify(sw));
-        break;
-      case 'a': // *s
-        const sa = _readUntrustedAnsi(arg);
-        a.push(JSON.stringify(sa));
-        break;
-      case 'Z': // *s[i]
-        const len = +args[j + 1];
-        const str = _readUntrustedUtf8(arg, len);
-        a.push(JSON.stringify(str));
-        break;
-      case 'S': // **s
-        const sss = _readUntrustedUtf8(Memory.readPointer(arg));
-        a.push(JSON.stringify(sss));
-        break;
-      case 'o':
-      case 'O':
-        if (ObjC.available) {
-          if (!arg.isNull()) {
-            if (darwin.isObjC(arg)) {
-              const o = new ObjC.Object(arg);
-              if (o.$className === 'Foundation.__NSSwiftData') {
-                a.push(`${o.$className}: "${ObjC.classes.NSString.alloc().initWithData_encoding_(o, 4).toString()}"`);
-              } else {
-                a.push(`${o.$className}: "${o.toString()}"`);
-              }
-            } else {
-              const str = Memory.readCString(arg);
-              if (str.length > 2) {
-                a.push(str);
-              } else {
-                a.push('' + arg);
-              }
-            }
-          } else {
-            a.push('nil');
-          }
-        } else {
-          a.push(arg);
-        }
-        break;
-      default:
-        a.push(arg);
-        break;
-    }
-  }
-  return { args: a, dumps: dumps };
-}
-
-function cloneArgs (args, fmt) {
-  const a = [];
-  let j = 0;
-  for (let i = 0; i < fmt.length; i++, j++) {
-    const arg = args[j];
-    switch (fmt[i]) {
-      case '+':
-      case '^':
-        j--;
-        break;
-      default:
-        a.push(arg);
-        break;
-    }
-  }
-  return a;
-}
-
-function _hexdumpUntrusted (addr, len) {
-  try {
-    if (typeof len === 'number') return hexdump(addr, { length: len });
-    else return hexdump(addr);
-  } catch (e) {
-    return `hexdump at ${addr} failed: ${e}`;
-  }
-}
-
-function _readUntrustedUtf8 (address, length) {
-  try {
-    if (typeof length === 'number') {
-      return Memory.readUtf8String(ptr(address), length);
-    }
-    return Memory.readUtf8String(ptr(address));
-  } catch (e) {
-    if (e.message !== 'invalid UTF-8') {
-      // TODO: just use this, doo not mess with utf8 imho
-      return Memory.readCString(ptr(address));
-    }
-    return '(invalid utf8)';
-  }
-}
-
-function _readUntrustedUtf16 (address, length) {
-  try {
-    if (typeof length === 'number') {
-      return Memory.readUtf16String(ptr(address), length);
-    }
-    return Memory.readUtf16String(ptr(address));
-  } catch (e) {
-    if (e.message !== 'invalid UTF-16') {
-      // TODO: just use this, doo not mess with utf8 imho
-      return Memory.readCString(ptr(address));
-    }
-    return '(invalid utf16)';
-  }
-}
-
-function _readUntrustedAnsi (address, length) {
-  try {
-    if (typeof length === 'number') {
-      return Memory.readAnsiString(ptr(address), length);
-    }
-    return Memory.readAnsiString(ptr(address));
-  } catch (e) {
-    if (e.message !== 'invalid Ansi') {
-      // TODO: just use this, doo not mess with utf8 imho
-      return Memory.readCString(ptr(address));
-    }
-    return '(invalid Ansi)';
-  }
-}
-
-function traceList () {
-  let count = 0;
-  return traceListeners.map((t) => {
-    return [count++, t.hits, t.at, t.source, t.moduleName, t.name, t.args].join('\t');
-  }).join('\n') + '\n';
-}
-
-function traceListJson () {
-  return traceListeners.map(_ => JSON.stringify(_)).join('\n') + '\n';
-}
-
-function traceHook (args) {
-  if (args.length === 0) {
-    return JSON.stringify(tracehooks, null, 2);
-  }
-  const arg = args[0];
-  if (arg !== undefined) {
-    tracehookSet(arg, args.slice(1).join(' '));
-  }
-  return '';
-}
-
-function traceFormat (args) {
-  if (args.length === 0) {
-    return traceList();
-  }
-  let address, format;
-  const name = args[0];
-  if (args.length === 2) {
-    address = '' + getPtr(name);
-    format = args[1];
-  } else if (args.length === 1) {
-    address = '' + getPtr(name);
-    format = '';
-  } else {
-    address = global.r2frida.offset;
-    format = args[0];
-  }
-  if (haveTraceAt(address)) {
-    return 'There\'s already a trace in here';
-  }
-  const traceOnEnter = format.indexOf('^') !== -1;
-  const traceBacktrace = format.indexOf('+') !== -1;
-  const useCmd = config.getString('hook.usecmd');
-
-  const currentModule = getModuleByAddress(address);
-  const listener = Interceptor.attach(ptr(address), {
-    myArgs: [],
-    myBacktrace: [],
-    keepArgs: [],
-    onEnter: function (args) {
-      traceListener.hits++;
-      if (!traceOnEnter) {
-        this.keepArgs = cloneArgs(args, format);
-      } else {
-        const fa = formatArgs(args, format);
-        this.myArgs = fa.args;
-        this.myDumps = fa.dumps;
-      }
-      if (traceBacktrace) {
-        this.myBacktrace = Thread.backtrace(this.context).map(DebugSymbol.fromAddress);
-      }
-      if (traceOnEnter) {
-        const traceMessage = {
-          source: 'dtf',
-          name: name,
-          address: address,
-          timestamp: new Date(),
-          values: this.myArgs,
-        };
-        if (config.getBoolean('hook.backtrace')) {
-          traceMessage.backtrace = Thread.backtrace(this.context).map(DebugSymbol.fromAddress);
-        }
-        if (config.getString('hook.output') === 'json') {
-          log.traceEmit(traceMessage);
-        } else {
-          let msg = `[dtf onEnter][${traceMessage.timestamp}] ${name}@${address} - args: ${this.myArgs.join(', ')}`;
-          if (config.getBoolean('hook.backtrace')) {
-            msg += ` backtrace: ${traceMessage.backtrace.toString()}`;
-          }
-          for (let i = 0; i < this.myDumps.length; i++) msg += `\ndump:${i + 1}\n${this.myDumps[i]}`;
-          log.traceEmit(msg);
-        }
-        if (useCmd.length > 0) {
-          console.log('[r2cmd]' + useCmd);
-        }
-      }
-    },
-    onLeave: function (retval) {
-      if (!traceOnEnter) {
-        const fa = formatArgs(this.keepArgs, format);
-        this.myArgs = fa.args;
-        this.myDumps = fa.dumps;
-
-        const traceMessage = {
-          source: 'dtf',
-          name: name,
-          address: address,
-          timestamp: new Date(),
-          values: this.myArgs,
-          retval
-        };
-        if (config.getBoolean('hook.backtrace')) {
-          traceMessage.backtrace = Thread.backtrace(this.context).map(DebugSymbol.fromAddress);
-        }
-        if (config.getString('hook.output') === 'json') {
-          log.traceEmit(traceMessage);
-        } else {
-          let msg = `[dtf onLeave][${traceMessage.timestamp}] ${name}@${address} - args: ${this.myArgs.join(', ')}. Retval: ${retval.toString()}`;
-          if (config.getBoolean('hook.backtrace')) {
-            msg += ` backtrace: ${traceMessage.backtrace.toString()}`;
-          }
-          for (let i = 0; i < this.myDumps.length; i++) msg += `\ndump:${i + 1}\n${this.myDumps[i]}`;
-          log.traceEmit(msg);
-        }
-        if (useCmd.length > 0) {
-          console.log('[r2cmd]' + useCmd);
-        }
-      }
-    }
-  });
-  const traceListener = {
-    source: 'dtf',
-    hits: 0,
-    at: ptr(address),
-    name: name,
-    moduleName: currentModule ? currentModule.name : '',
-    format: format,
-    listener: listener
-  };
-  traceListeners.push(traceListener);
-  return true;
-}
-
-function traceListenerFromAddress (address) {
-  const results = traceListeners.filter((tl) => '' + address === '' + tl.at);
-  return (results.length > 0) ? results[0] : undefined;
-}
-
-function traceCountFromAddress (address) {
-  const tl = traceListenerFromAddress(address);
-  return tl ? tl.hits : 0;
-}
-
-function traceNameFromAddress (address) {
-  const tl = traceListenerFromAddress(address);
-  return tl ? tl.moduleName + ':' + tl.name : '';
-}
-
-function traceLogDumpQuiet () {
-  return log.logs.map(({ address, timestamp }) =>
-    [address, timestamp, traceCountFromAddress(address), traceNameFromAddress(address)].join(' '))
-    .join('\n') + '\n';
-}
-
-function traceLogDumpJson () {
-  return JSON.stringify(log.logs);
-}
-
-function traceLogDumpR2 () {
-  let res = '';
-  for (const l of log.logs) {
-    if (l.script) {
-      res += l.script;
-    }
-  }
-  return res;
-}
-
-function objectToString (o) {
-  // console.error(JSON.stringify(o));
-  const r = Object.keys(o).map((k) => {
-    try {
-      const p = ptr(o[k]);
-      if (darwin.isObjC(p)) {
-        const o = new ObjC.Object(p);
-        return k + ': ' + o.toString();
-      }
-      const str = Memory.readCString(p);
-      if (str.length > 2) {
-        return k + ': "' + str + '"';
-      }
-    } catch (e) {
-    }
-    return k + ': ' + o[k];
-  }).join(' ');
-  return '(' + r + ')';
-}
-
-function tracelogToString (l) {
-  const line = [l.source, l.name || l.address, objectToString(l.values)].join('\t');
-  const bt = (!l.backtrace)
-    ? ''
-    : l.backtrace.map((b) => {
-      return ['', b.address, b.moduleName, b.name].join('\t');
-    }).join('\n') + '\n';
-  return line + bt;
-}
-
-function traceLogDump () {
-  return log.logs.map(tracelogToString).join('\n') + '\n';
-}
-
-function traceLogClear (args) {
-  // TODO: clear one trace instead of all
-  console.error('ARGS', JSON.stringify(args));
-  return traceLogClearAll();
-}
-
-function traceLogClearAll () {
-  log.logs = [];
-  log.traces = {};
-  return '';
-}
-
-
-
-function haveTraceAt (address) {
-  try {
-    for (const trace of traceListeners) {
-      if (trace.at.compare(address) === 0) {
-        return true;
-      }
-    }
-  } catch (e) {
-    console.error(e);
-  }
-  return false;
-}
-
-function traceRegs (args) {
-  if (args.length < 1) {
-    return 'Usage: dtr [name|address] [reg ...]';
-  }
-  const address = getPtr(args[0]);
-  if (haveTraceAt(address)) {
-    return 'There\'s already a trace in here';
-  }
-  const rest = args.slice(1);
-  const currentModule = getModuleByAddress(address);
-  const listener = Interceptor.attach(address, traceFunction);
-  function traceFunction (_) {
-    traceListener.hits++;
-    const regState = {};
-    rest.forEach((r) => {
-      let regName = r;
-      let regValue;
-      if (r.indexOf('=') !== -1) {
-        const kv = r.split('=');
-        this.context[kv[0]] = ptr(kv[1]); // set register value
-        regName = kv[0];
-        regValue = kv[1];
-      } else {
-        try {
-          const rv = ptr(this.context[r]);
-          regValue = rv;
-          let tail = Memory.readCString(rv);
-          if (tail) {
-            tail = ' (' + tail + ')';
-            regValue += tail;
-          }
-        } catch (e) {
-          // do nothing
-        }
-      }
-      regState[regName] = regValue;
-    });
-    const traceMessage = {
-      source: 'dtr',
-      address: address,
-      timestamp: new Date(),
-      values: regState,
-    };
-    if (config.getBoolean('hook.backtrace')) {
-      traceMessage.backtrace = Thread.backtrace(this.context).map(DebugSymbol.fromAddress);
-    }
-    if (config.getString('hook.output') === 'json') {
-      log.traceEmit(traceMessage);
-    } else {
-      let msg = `[dtr][${traceMessage.timestamp}] ${address} - registers: ${JSON.stringify(regState)}`;
-      if (config.getBoolean('hook.backtrace')) {
-        msg += ` backtrace: ${traceMessage.backtrace.toString()}`;
-      }
-      log.traceEmit(msg);
-    }
-  }
-  const traceListener = {
-    source: 'dtr',
-    hits: 0,
-    at: address,
-    moduleName: currentModule ? currentModule.name : 'unknown',
-    name: args[0],
-    listener: listener,
-    args: rest
-  };
-  traceListeners.push(traceListener);
-  return '';
-}
-
-function traceHere () {
-  const args = [r2frida.offset];
-  args.forEach(address => {
-    const at = DebugSymbol.fromAddress(ptr(address)) || '' + ptr(address);
-    const listener = Interceptor.attach(ptr(address), function () {
-      const bt = Thread.backtrace(this.context).map(DebugSymbol.fromAddress);
-      const at = nameFromAddress(address);
-      console.log('Trace here probe hit at ' + address + '::' + at + '\n\t' + bt.join('\n\t'));
-    });
-    traceListeners.push({
-      at: at,
-      listener: listener
-    });
-  });
-  return true;
-}
-
-function traceR2 (args) {
-  return traceListeners.map(_ => `dt+ ${_.at} ${_.hits}`).join('\n') + '\n';
-}
-
 function dumpJavaArguments (args) {
   let res = '';
   try {
@@ -1102,87 +596,6 @@ function traceJavaConstructors (className) {
         return over.apply(this, arguments);
       };
     });
-  });
-}
-
-function traceSwift (klass, method) {
-  if (!swift.SwiftAvailable()) {
-    return;
-  }
-  const targetAddress = getPtr('swift:' + klass + '.' + method);
-  if (ptr(0).equals(targetAddress)) {
-    console.error('Missing method ' + method + ' in class ' + klass);
-    return;
-  }
-
-  const callback = function (args) {
-    const msg = ['[SWIFT]', klass, method, JSON.stringify(args)];
-    log.traceEmit(msg.join(' '));
-  };
-  Swift.Interceptor.Attach(target, callback);
-}
-
-function traceJava (klass, method) {
-  javaPerform(function () {
-    const Throwable = Java.use('java.lang.Throwable');
-    const k = javaUse(klass);
-    k[method].implementation = function (args) {
-      const res = this[method]();
-      const bt = config.getBoolean('hook.backtrace')
-        ? Throwable.$new().getStackTrace().map(_ => _.toString())
-        : [];
-      const traceMessage = {
-        source: 'dt',
-        klass: klass,
-        method: method,
-        backtrace: bt,
-        timestamp: new Date(),
-        result: res,
-        values: args
-      };
-      if (config.getString('hook.output') === 'json') {
-        log.traceEmit(traceMessage);
-      } else {
-        let msg = `[JAVA TRACE][${traceMessage.timestamp}] ${klass}:${method} - args: ${JSON.stringify(args)}. Return value: ${res.toString()}`;
-        if (config.getBoolean('hook.backtrace')) {
-          msg += ` backtrace: \n${traceMessage.backtrace.toString().split(',').join('\nat ')}\n`;
-        }
-        log.traceEmit(msg);
-      }
-      return res;
-    };
-  });
-}
-
-function traceQuiet (args) {
-  return traceListeners.map(({ address, hits, moduleName, name }) => [address, hits, moduleName + ':' + name].join(' ')).join('\n') + '\n';
-}
-
-function traceJson (args) {
-  if (args.length === 0) {
-    return traceListJson();
-  }
-  if (args[0].startsWith('java:')) {
-    traceReal(args[0]);
-    return;
-  }
-  return new Promise(function (resolve, reject) {
-    (function pull () {
-      const arg = args.pop();
-      if (arg === undefined) {
-        return resolve('');
-      }
-      const narg = getPtr(arg);
-      if (narg) {
-        traceReal(arg, narg);
-        pull();
-      } else {
-        numEval(arg).then(function (at) {
-          console.error(traceReal(arg, at));
-          pull();
-        }).catch(reject);
-      }
-    })();
   });
 }
 
@@ -1345,169 +758,6 @@ function swiftTypes (args) {
       break;
   }
   return res;
-}
-
-function trace (args) {
-  if (args.length === 0) {
-    return traceList();
-  }
-  return traceJson(args);
-}
-
-function tracehookSet (name, format, callback) {
-  if (name === null) {
-    console.error('Name was not resolved');
-    return false;
-  }
-  tracehooks[name] = {
-    format: format,
-    callback: callback
-  };
-  return true;
-}
-
-function arrayBufferToHex (arrayBuffer) {
-  if (typeof arrayBuffer !== 'object' || arrayBuffer === null || typeof arrayBuffer.byteLength !== 'number') {
-    throw new TypeError('Expected input to be an ArrayBuffer');
-  }
-
-  const view = new Uint8Array(arrayBuffer);
-  let result = '';
-  let value;
-
-  for (let i = 0; i < view.length; i++) {
-    value = view[i].toString(16);
-    result += (value.length === 1 ? '0' + value : value);
-  }
-
-  return result;
-}
-
-// \dth printf 0,1 .. kind of dtf
-function tracehook (address, args) {
-  const at = nameFromAddress(address);
-  const th = tracehooks[at];
-  const fmtarg = [];
-  if (th && th.format) {
-    for (const fmt of th.format.split(' ')) {
-      const [k, v] = fmt.split(':');
-      switch (k) {
-        case 'i':
-        // console.log('int', args[v]);
-          fmtarg.push(+args[v]);
-          break;
-        case 's':
-          {
-            const [a, l] = v.split(',');
-            const addr = ptr(args[a]);
-            const size = +args[l];
-            // const buf = Memory.readByteArray(addr, size);
-            // console.log('buf', arrayBufferToHex(buf));
-            // console.log('string', Memory.readCString(addr, size));
-            fmtarg.push(Memory.readCString(addr, size));
-          }
-          break;
-        case 'z':
-        // console.log('string', Memory.readCString(args[+v]));
-          fmtarg.push(Memory.readCString(ptr(args[+v])));
-          break;
-        case 'v':
-          {
-            const [a, l] = v.split(',');
-            const addr = ptr(args[a]);
-            const buf = Memory.readByteArray(addr, +args[l]);
-            fmtarg.push(arrayBufferToHex(buf));
-          }
-          break;
-      }
-    }
-  }
-  return fmtarg;
-}
-
-function traceReal (name, addressString) {
-  if (arguments.length === 0) {
-    return traceList();
-  }
-  if (name.startsWith('swift:')) {
-    const km = name.substring(6);
-    const dot = km.lastIndexOf('.');
-    if (dot === -1) {
-      return 'Invalid syntax for swift uri. Use "swift:KLASS.METHOD"';
-    }
-    const klass = km.substring(0, dot);
-    const methd = km.substring(dot + 1);
-    return traceSwift(klass, methd);
-  }
-  if (name.startsWith('java:')) {
-    const javaName = name.substring(5);
-    if (javaUse(javaName)) {
-      console.error('Tracing class constructors');
-      traceJavaConstructors(javaName);
-    } else {
-      const dot = javaName.lastIndexOf('.');
-      if (dot !== -1) {
-        const klass = javaName.substring(0, dot);
-        const methd = javaName.substring(dot + 1);
-        traceJava(klass, methd);
-      } else {
-        console.log('Invalid java method name. Use :dt java:package.class.method');
-      }
-    }
-    return;
-  }
-  const address = ptr(addressString);
-  if (haveTraceAt(address)) {
-    return 'There\'s already a trace in here';
-  }
-  const currentModule = getModuleByAddress(address);
-  const listener = Interceptor.attach(address, function (args) {
-    const values = tracehook(address, args);
-    const traceMessage = {
-      source: 'dt',
-      address: address,
-      timestamp: new Date(),
-      values: values,
-    };
-    traceListener.hits++;
-    if (config.getString('hook.output') === 'json') {
-      log.traceEmit(traceMessage);
-    } else {
-      log.traceEmit(`[dt][${traceMessage.timestamp}] ${address} - args: ${JSON.stringify(values)}`);
-    }
-  });
-  const traceListener = {
-    source: 'dt',
-    at: address,
-    hits: 0,
-    name: name,
-    moduleName: currentModule ? currentModule.name : 'unknown',
-    args: '',
-    listener: listener
-  };
-  traceListeners.push(traceListener);
-  return '';
-}
-
-function clearAllTrace (args) {
-  traceListeners.splice(0).forEach(lo => lo.listener ? lo.listener.detach() : null);
-  return '';
-}
-
-function clearTrace (args) {
-  if (args.length > 0 && +args[0] > 0) {
-    const res = [];
-    const nth = +args[0];
-    for (let i = 0; i < traceListeners.length; i++) {
-      const tl = traceListeners[i];
-      if (i === nth) {
-        tl.listener.detach();
-      } else {
-        res.push(tl);
-      }
-    }
-  }
-  return '';
 }
 
 function interceptHelp (args) {
